Extract test helper method to separate place
